+++ backend/de.metas.cucumber/src/test/java/de/metas/cucumber/stepdefs/accounting/FactAcctMatchersFactory.java
@@ -5,7 +5,7 @@ import com.google.common.collect.ImmutableListMultimap;
 import de.metas.acct.AccountConceptualName;
 import de.metas.bpartner.BPartnerId;
 import de.metas.cucumber.stepdefs.C_BPartner_StepDefData;
-import de.metas.cucumber.stepdefs.C_Tax_StepDefData;
+import de.metas.cucumber.stepdefs.tax.C_Tax_StepDefData;
 import de.metas.cucumber.stepdefs.DataTableRow;
 import de.metas.cucumber.stepdefs.DataTableRows;
 import de.metas.cucumber.stepdefs.M_Product_StepDefData;
