+++ backend/de.metas.cucumber/src/test/java/de/metas/cucumber/stepdefs/accounting/Fact_Acct_StepDef.java
@@ -2,7 +2,7 @@ package de.metas.cucumber.stepdefs.accounting;
 
 import com.google.common.collect.ImmutableSet;
 import de.metas.cucumber.stepdefs.C_BPartner_StepDefData;
-import de.metas.cucumber.stepdefs.C_Tax_StepDefData;
+import de.metas.cucumber.stepdefs.tax.C_Tax_StepDefData;
 import de.metas.cucumber.stepdefs.M_Product_StepDefData;
 import de.metas.cucumber.stepdefs.invoice.C_Invoice_StepDefData;
 import de.metas.cucumber.stepdefs.util.IdentifiersResolver;
