+++ backend/de.metas.cucumber/src/test/java/de/metas/cucumber/stepdefs/invoice/C_InvoiceTax_StepDef.java
@@ -23,7 +23,7 @@
 package de.metas.cucumber.stepdefs.invoice;
 
 import com.google.common.collect.ImmutableList;
-import de.metas.cucumber.stepdefs.C_Tax_StepDefData;
+import de.metas.cucumber.stepdefs.tax.C_Tax_StepDefData;
 import de.metas.cucumber.stepdefs.DataTableRow;
 import de.metas.cucumber.stepdefs.DataTableRows;
 import de.metas.cucumber.stepdefs.StepDefDataIdentifier;
