+++ backend/de.metas.cucumber/src/test/java/de/metas/cucumber/stepdefs/pricing/M_Product_TaxCategory_StepDef.java
@@ -25,6 +25,7 @@ package de.metas.cucumber.stepdefs.pricing;
 import de.metas.cucumber.stepdefs.C_Country_StepDefData;
 import de.metas.cucumber.stepdefs.DataTableUtil;
 import de.metas.cucumber.stepdefs.M_Product_StepDefData;
+import de.metas.cucumber.stepdefs.tax.C_TaxCategory_StepDefData;
 import de.metas.util.Services;
 import io.cucumber.datatable.DataTable;
 import io.cucumber.java.en.Given;
