+++ backend/de.metas.cucumber/src/test/java/de/metas/cucumber/stepdefs/tax/C_TaxCategory_StepDefData.java
@@ -20,7 +20,7 @@
  * #L%
  */
 
-package de.metas.cucumber.stepdefs.pricing;
+package de.metas.cucumber.stepdefs.tax;
 
 import de.metas.cucumber.stepdefs.StepDefData;
 import de.metas.cucumber.stepdefs.StepDefDataGetIdAware;
