+++ backend/de.metas.cucumber/src/test/java/de/metas/cucumber/stepdefs/tax/C_Tax_StepDefData.java
@@ -20,8 +20,10 @@
  * #L%
  */
 
-package de.metas.cucumber.stepdefs;
+package de.metas.cucumber.stepdefs.tax;
 
+import de.metas.cucumber.stepdefs.StepDefData;
+import de.metas.cucumber.stepdefs.StepDefDataGetIdAware;
 import de.metas.tax.api.Tax;
 import de.metas.tax.api.TaxId;
