+++ backend/de.metas.cucumber/src/test/java/de/metas/cucumber/stepdefs/tax/C_VAT_Code_StepDef.java
@@ -26,7 +26,6 @@ import de.metas.acct.api.IAcctSchemaBL;
 import de.metas.acct.vatcode.CreateVATCodeRequest;
 import de.metas.acct.vatcode.IVATCodeDAO;
 import de.metas.acct.vatcode.VATCode;
-import de.metas.cucumber.stepdefs.C_Tax_StepDefData;
 import de.metas.cucumber.stepdefs.DataTableRow;
 import de.metas.cucumber.stepdefs.DataTableRows;
 import de.metas.cucumber.stepdefs.StepDefConstants;
